Refactor because load_openai was moved to controller package
