+++ controller/stream_completion.py
@@ -23,7 +23,7 @@ from elevenlabs import generate, stream, save
 from elevenlabs.api import Voice, VoiceSettings
 import pyaudio
 
-from controller.waifuai.load_openai import load_openai
+from controller.load_openai import load_openai
 from controller.create_logger import create_logger
 from controller.waifuai.record_voice import save_wav_file
