+++ controller/utilities.py
@@ -15,7 +15,7 @@ from tenacity import retry, wait_random_exponential, stop_after_attempt
 from termcolor import colored
 
 from controller.create_logger import create_logger
-from controller.waifuai.load_openai import load_openai
+from controller.load_openai import load_openai
 
 # Create a logger for this module
 module_logger = create_logger(
