+++ controller/waifuai/completion_create.py
@@ -5,7 +5,7 @@ import json
 
 import openai
 
-from controller.waifuai.load_openai import load_openai
+from controller.load_openai import load_openai
 from controller.create_logger import create_logger
 from controller.ai_functions import available_functions
